UIBULKED-415 Refactoring of bulk-edit preview components
